Move shell to separate file
